Moving core related global functions to a separate file
